requirements.txt
asgiref==3.7.1
attrs==23.1.0
distlib==0.3.6
Django==4.2.1
django-cors-headers==4.1.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.2
environ==1.0
filelock==3.12.0
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.5.0
platformdirs==3.5.1
psycopg2==2.9.6
PyJWT==2.8.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
